Add autotools for building the client
This enables easier building for different targets, e.g. LEDE/OpenWrt or
other Linux distributions.

Also enables building with shared libasyncns, which is turned on by
default if the shared version is found on the system.
For this the libasyncns source files were moved to their own
subdirectory so that the include statement will pull in the header file
for the shared version instead of the bundled one.

I also made the decision to check in all required generated files and
not include them in the gitignore file. This should make releasing and
building releases much easier. However, this means that changes in
either configure.ac or Makefile.am require an autoreconf and subsequent
check-in of the newly generated files.

Signed-off-by: Felix Kaechele <felix@kaechele.ca>