client: Switch libnl-tiny and libnl in cmake
This issue unfortunately popped up after I commited the cmake file.

In LEDE/OpenWrt buildroots cmake will find the full libnl first if both
are available in the buildroot.

Switching the two alternatives will give preference to libnl-tiny, which
is what most likely will end up in the image.

Also this introduces a flag to force which libnl implementation is to be
used. This allows for more control over the build process while
maintaining the convenience of autodetection in most other environments.

A few style fixes are included as well, namely changing all command
invocations to lowercase consistently throughout the file.

Signed-off-by: Felix Kaechele <felix@kaechele.ca>