client: Implement IPv6
This adds IPv6 capabilities to the client. A fallback to IPv4 has been
implemented for DNS names that return both an A and AAAA record but
where the broker only replies to IPv4 requests.

This removes binding to local IPs or interfaces as we connect() the
socket which makes the Kernel choose the local endpoint automatically
based on the interface that was used for the outgoing request.

IPv6 Addresses can be passed to the client using the -b parameter like
this:

  tunneldigger [opts] -b [2001:db8::1]:8942

you can also specify the interface scope for IPv6 like this:

  tunneldigger [opts] -b [fe80::1%ens3]:8942

Signed-off-by: Felix Kaechele <felix@kaechele.ca>