* made interface to an optional parameter
Since we are now running a tunneligger broker instance on a routed IP address, I noticed that the tunneligger does not work with a loopback interface.

To the basic structure:
Our IPv4 subnet is announced via BGP on several interfaces. An IP from the subnet is configured on the loopback interface.

A short analysis of the pollable object showed that the UDP socket is bound to an interface.  I could not find out why this was done.

This change allows to define interfaces optionally.